Rename ecosystem.d.ts to ecosystemTypes.ts
This allows importing Enums so they don't have to be redefined in the
mod manager.
