+++ src/assets/data/ecosystemTypes.ts
