+++ src/model/schema/ThunderstoreSchema.ts
@@ -1,8 +1,8 @@
 import Ajv from "ajv";
 import addFormats from "ajv-formats";
 
-import { ThunderstoreEcosystem }  from "../../assets/data/ecosystem.d";
 import ecosystem from "../../assets/data/ecosystem.json";
+import { ThunderstoreEcosystem }  from "../../assets/data/ecosystemTypes";
 import jsonSchema from "../../assets/data/ecosystemJsonSchema.json";
 import R2Error from "../errors/R2Error";
